+++ tests/reflection/global-uniforms.hlsl
@@ -1,4 +1,4 @@
-//TEST:SIMPLE:-profile ps_4_0 -target hlsl -target reflection-json
+//TEST:REFLECTION:-profile ps_4_0 -target hlsl
 
 // Confirm that we handle uniforms at global scope
