+++ tests/reflection/thread-group-size.comp
@@ -1,4 +1,4 @@
-//TEST:SIMPLE:-no-checking -target glsl -target reflection-json
+//TEST:REFLECTION:-no-checking -target glsl
 
 // Confirm that we provide reflection data for the `local_size_*` attributes
